+++ monitoring-common/pom.xml
@@ -24,7 +24,7 @@
     <version>3.5.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>vertx-monitoring-base</artifactId>
+  <artifactId>vertx-monitoring-common</artifactId>
 
   <name>Vert.x Monitoring Base</name>
