+++ monitoring-common/src/main/asciidoc/dataobjects.adoc
