+++ monitoring-common/src/main/asciidoc/enums.adoc
