+++ monitoring-common/src/main/java/io/vertx/ext/monitoring/common/MetricsOptionsBase.java
@@ -13,7 +13,7 @@
  *
  * You may elect to redistribute this code under either of these licenses.
  */
-package io.vertx.ext.monitoring.base;
+package io.vertx.ext.monitoring.common;
 
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.codegen.annotations.GenIgnore;
